Change a litte the magic array

+++ Foundation/magic_array_public.h
@@ -25,7 +25,7 @@ void magicArrayClear(MAGIC_ARRAY* magicArray);
 unsigned int magicArrayGetLength(MAGIC_ARRAY* magicArray);
 void* magicArrayAddItem(MAGIC_ARRAY* magicArray);
 void* magicArrayInsertItem(MAGIC_ARRAY* magicArray, unsigned int itemIndex);
-void magicArrayRemoveItemByIndex(MAGIC_ARRAY* magicArray, unsigned int itemIndex);
+void magicArrayRemoveItemByIndex(MAGIC_ARRAY* magicArray, unsigned long itemIndex);
 int magicArrayRemoveItem(MAGIC_ARRAY* magicArray, void* item);
 void* magicArrayGetItem(MAGIC_ARRAY* magicArray, unsigned long itemIndex);
 unsigned int magicArrayGetItemIndex(MAGIC_ARRAY* array, void* item);
